REFATORADO CRUD de Revista

+++ database/migrations/2021_11_30_004219_artigo_autor.php
@@ -15,8 +15,8 @@ class ArtigoAutor extends Migration
     {
         Schema::create('ArtigoAutor', function (Blueprint $table) {
 
-            $table->integer('artigo_id')->unsigned();        
-            $table->integer('autor_id')->unsigned();
+            $table->bigInteger('artigo_id')->unsigned();        
+            $table->bigInteger('autor_id')->unsigned();
             
             $table->foreign('artigo_id')->references('id')->on('artigos')->onDelete('cascade');
             $table->foreign('autor_id')->references('id')->on('autors')->onDelete('cascade');
